+++ server/src-lib/Hasura/GraphQL/Schema/Select.hs
@@ -41,8 +41,8 @@ import qualified Hasura.Backends.Postgres.SQL.DML      as SQL
 import qualified Hasura.GraphQL.Execute.Types          as ET
 import qualified Hasura.GraphQL.Parser                 as P
 import qualified Hasura.GraphQL.Parser.Internal.Parser as P
+import qualified Hasura.RQL.IR.BoolExp                 as RQL
 import qualified Hasura.RQL.IR.Select                  as RQL
-import qualified Hasura.RQL.Types.BoolExp              as RQL
 
 import           Data.Text.Extended
 import           Hasura.Backends.Postgres.SQL.Types
