+++ server/src-lib/Hasura/RQL/IR/Delete.hs
@@ -5,8 +5,8 @@ import           Hasura.Prelude
 import qualified Hasura.Backends.Postgres.SQL.DML   as S
 
 import           Hasura.Backends.Postgres.SQL.Types
+import           Hasura.RQL.IR.BoolExp
 import           Hasura.RQL.IR.Returning
-import           Hasura.RQL.Types.BoolExp
 import           Hasura.RQL.Types.Column
 import           Hasura.SQL.Backend
