+++ server/src-lib/Hasura/RQL/IR/Select.hs
@@ -17,7 +17,7 @@ import qualified Hasura.Backends.Postgres.SQL.DML    as S
 
 import           Hasura.Backends.Postgres.SQL.Types
 import           Hasura.GraphQL.Parser.Schema
-import           Hasura.RQL.Types.BoolExp
+import           Hasura.RQL.IR.BoolExp
 import           Hasura.RQL.Types.Column
 import           Hasura.RQL.Types.Common
 import           Hasura.RQL.Types.DML
